Crop the images tightly.
The problem. When displayed math is used (e.g. $$math$$), there is empty space on images at the left (PNG) and right (PDF) sides.
https://tex.stackexchange.com/q/13981/222496

The cause is not well known. The 'preview' mode doesn't crop the images horizontally. So the 'crop' option must be used for the 'standalone' package instead of 'preview'. But when we use 'crop', displayed math doesn't work (https://tex.stackexchange.com/a/50171/222496) and we must use the 'varwidth' option to make it work. But when we use the 'varwidth' option, the images are tightly cropped even with the 'preview' option.

The solution. Add the 'varwidth' option that makes cropping tight.
We use 10000pt width to widen PDF page so that it can fit large enough equations (16383 is maximal allowed width). It's only for PDF. PNG images are extended to fit equations of any width.
Also we use the 'border' option. The LaTeX content may be clipped in PDF without this option. PNG images doesn't need this option.