Add the DIRECT_EMBED option.
The problem. Pages that contain latex macros are broken on old Redmine versions. The title is not displayed. And there is no right side panel.

The cause. The root cause is unknown. The problem happens when calling rendering methods of controller.

The solution. Do not call rendering methods. Render directly in the 'Macro' class code without using templates.